pyproject.toml
[project]
name = "rag-que"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "google-cloud-storage>=3.1.0",
    "openai>=1.84.0",
    "pyjwt>=2.10.1",
    "pymilvus>=2.5.10",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "uvicorn>=0.34.3",
]
